add base config
